Env installation and infrastructure created
I have created the first infrastructure files of the server that create a connection with mongoDB provided by the env file and then globalizes the infrastructure to make it accesible around the app.

Also i have created all the env configuration (not that much) and changed the port to serve to the one specified in the env file.

An example of that env was created at .env-example