Pets helpers and re-organize the start of the application
I have created all the CRUD pets helpers for the application.

Also i have re-created the entry point of the application to wait to start the infrastructure to start all the others services because any service needs to access to the infrastructure since the first time and i let the same globalization of some components but using an async function.